Updated Coordinator to use new messaging method
New method involves a stream buffer feeding an interrupt to have interrupt driven outputs that aren't reliant on task priority at all. This should be faster, and less likely to drop out in the middle of a message.

Also implemented a wired network init routine that generates a connected device table for the round robin task.