indicator updates complete?
indicator updates are complete for now, error indication not yet implemented but it shouldn't be too hard to add a case for it.

I will work on reviewing code to catch obvious mistakes before I go to single device testing. I will test by feeding messages to the device to ensure it reacts properly