copied 485SecControl
I copied the 485SecControl over as a test. I want to be sure I can use it as a template to save time when making the other devices.

After a build test, I discovered that I need to define F_CPU somewhere that it can be seen by the functions using it.