wired comms optimization work
Wired comms is functional, but it starts to bog down as more devices are added to a channel. I will try to reduce this by ensuring that only a single message needs to be sent on the wired network per command. The issue is somewhat unavoidable with many distinct wireless devices, but at least it won't be a problem on the wired network.